+++ changes.md
@@ -1,5 +1,6 @@
 ##### unreleased
 
+-   Added flash messages for user feedback in admin. [#17]
 -   Added punishment points entries editing in admin. [#16]
 -   Added race result entries editing in admin. [#15]
 -   Added punishment points admin. [#14]